Task-2/requirements.txt
async-timeout==4.0.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
paho-mqtt==1.6.1
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
redis==5.0.1
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
Werkzeug==3.0.1
